Issue 55: Split up large file.
Rename classes to match their functionality more closely.
