+++ src/pytestarch/rule_assessment/rule_check/rule_violations.py
@@ -1,0 +1,29 @@
+from dataclasses import dataclass, fields
+from typing import List, Optional
+
+from pytestarch.eval_structure.evaluable_architecture import (
+    LaxDependenciesByBaseModule,
+    StrictDependenciesByBaseModules,
+)
+
+
+@dataclass
+class RuleViolations:
+    lax_dependencies: Optional[LaxDependenciesByBaseModule]
+    strict_dependencies: Optional[StrictDependenciesByBaseModules]
+
+    should_violated: bool = False
+    should_only_violated_by_forbidden_import: bool = False
+    should_only_violated_by_no_import: bool = False
+    should_not_violated: bool = False
+    should_except_violated: bool = False
+    should_only_except_violated_by_forbidden_import: bool = False
+    should_only_except_violated_by_no_import: bool = False
+    should_not_except_violated: bool = False
+
+    def __bool__(self) -> bool:
+        return any(self.__dict__[a] for a in self._bool_field_names())
+
+    @classmethod
+    def _bool_field_names(cls) -> List[str]:
+        return [field.name for field in fields(cls) if field.type == bool]
